Copy/paste modifies a view aggregator for that directory
        This copies the templates directory scrapping code and modifies
it. It works, but there is a fair amount of duplication. I took 10
minutes to try to aggregate the code and make it more configurable with
require vs read and extension names. It got a lot uglier and harder to
understand and so I gave up.

        If this code duplication bites anyone, then probably some little
extractions would be helpful.